Adds a guard on release finish to ensure latest changes
This will now ask developers if they want to update their local release
branch before proceeding if there have been changes made on the remote.

If the developer specifies 'n' the remote changes will be thrown away
and changes will be force pushed. This ensures that the developer is
releasing everything that they know that has been verified.

If the developer specifies 'y' the remote changes will be fetched /
pulled in and the release finish process will continue as normal.
Developers should be wary of doing this, and should get in the habit of
pulling before finishing a release.